代码构建迭代完成
